Add guards for when a server timer controls state
When the server controls the location state (e.g. Forbidden Grove, Balack's Cove, Seasonal Garden, and Gnawnian Express Station), there is a possibility of the state changing between when the user object was generated pre-hunt, and when the hunt occurs.

For cases in which the user cannot be auto-traveled by a mouse catch (e.g. Seasonal Garden & GES), require details of the pre- & post-hunt user objects to agree.
For cases in which the user can be auto-traveled by a mouse catch, the post-hunt user object may not contain the needed information. Use provided progress information to determine if a location state change is imminent, and avoid tracking if so.