Fix parsing the amount of loot when >999
The regex for parsing the amount of loot had a group around the
thousands separator and subsequent digits. This caused the match
call to return the base match and the group result. After replacing
commas with nothing, this resulted in a very large number.

This was only an issue for loot items that can drop in amounts of 1000
or more, like Gold or Cursed Gold.